fix(projection-art): eval-fix — 씬 컴포넌트 분리 + R3F 경고 제거 + 미사용 타입 삭제
RF-1: AudioReactiveVisual.tsx 632줄 → 340줄
  - MusicScene.tsx: WaveformLine, FrequencyBars, AudioParticles, MusicScene
  - VoiceScene.tsx: VoiceWave, EnergyRing, VoiceScene

RF-2: setup.ts에 R3F jsdom 경고 필터 추가
  - toneMapped, itemSize, wireframe 등 R3F 전용 prop 경고 억제

RF-3: types/index.ts에서 미사용 AudioAnalyzerState 인터페이스 제거
